# Conflicts:
#	include/dto/sqs/SQSMessageAttributeDataType.h

+++ include/utils/StringUtils.h
@@ -80,7 +80,7 @@ public:
 
     static QString FormatSizeColumn(const long value, const int precision) {
         if (value < 1024) {
-            return QString::number(static_cast<double>(value), 'f', 0) + " b";
+            return QString::number(value, 'f', 0) + " b";
         }
         if (value < 1024 * 1024) {
             return QString::number(static_cast<double>(value) / 1024, 'f', precision) + " kb";
